Added optional args and type checking to edgeWeights.
The signature of the edgeWeights generic is now:
setGeneric("edgeWeights", function(object, index, ...,
                                   attr="weight", default=1,
                                   type.checker=is.numeric)

The optional arguments control the name of the edge attribute 
used to extract weight values, the default value to use if there
is no edge attribute of the specified name defined for the graph, 
and the function used to verify that any extracted weights are 
of the expected type.

NOTE: By default, the expected type of edge weight values is numeric.
If edgeWeights is called on a graph with non-numeric edge weights,
an error will be signaled.  This can be avoided by passing an 
appropriate function via type.checker or specifying NULL to avoid
all type checking entirely.



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/graph@18227 bc3139a8-67e5-0310-9ffc-ced21a209358